# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: api
  description: ''
  license:
    name: ''
  version: 0.1.0
servers:
- url: /api
paths:
...
  /auth/current:
    get:
      tags:
      - auth
      operationId: current
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                email: alex@example.com
                is_admin: false
                name: Alex Rider
                pid: a1b2c3d4-0000-4000-8000-000000000012
                verified: true
        '401':
          description: Not authenticated
      security:
      - bearer_auth: []
...
  /strava/connect:
    post:
      tags:
      - strava
      operationId: begin_connect
      responses:
        '200':
          description: Strava authorization URL for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StravaAuthorizeResponse'
              example:
                authorization_url: https://www.strava.com/oauth/authorize?client_id=12345&response_type=code
        '400':
          description: Strava integration is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Not authenticated
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
  /strava/connection:
    get:
      tags:
      - strava
      operationId: get_connection
      responses:
        '200':
          description: Current Strava connection state for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StravaConnectionResponse'
              example:
                athlete_id: 12345678
                athlete_name: Alex Rider
                athlete_username: alex-rider
                configured: true
                connected: true
                last_sync_duplicate_count: 1
                last_sync_failed_count: 0
                last_sync_finished_at: 2026-09-27T12:00:00Z
                last_sync_imported_count: 3
                last_sync_status: succeeded
                scopes:
                - read
                - activity:read_all
        '401':
          description: Not authenticated
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
    delete:
      tags:
      - strava
      operationId: disconnect_connection
      responses:
        '200':
          description: Removed the authenticated user's Strava connection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Strava connection removed.
        '401':
          description: Not authenticated
        '409':
          description: The user's Strava sync is queued or running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
  /strava/sync:
    post:
      tags:
      - strava
      operationId: queue_sync
      responses:
        '200':
          description: Queued a Strava sync for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StravaConnectionResponse'
              example:
                athlete_id: 12345678
                athlete_name: Alex Rider
                configured: true
                connected: true
                last_sync_duplicate_count: 1
                last_sync_failed_count: 0
                last_sync_imported_count: 3
                last_sync_status: queued
                scopes:
                - read
                - activity:read_all
        '400':
          description: Strava integration is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Not authenticated
        '404':
          description: No Strava connection exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Another activity import is already running or queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
  /strava/webhook:
    get:
      tags:
      - strava
      operationId: handle_webhook_verification
      responses:
        '200':
          description: Verifies the Strava webhook subscription handshake
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StravaWebhookChallengeResponse'
              example:
                hub.challenge: strava-verification-challenge
        '400':
          description: Invalid handshake query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags:
      - strava
      operationId: handle_webhook_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StravaWebhookEvent'
            example:
              aspect_type: create
              event_time: 1790500000
              object_id: 1234567890
              object_type: activity
              owner_id: 12345678
              subscription_id: 12345
              updates: {}
        required: true
      responses:
        '200':
          description: Accepted a Strava webhook event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: ok
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      required:
      - message
      properties:
        errors:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
          propertyNames:
            type: string
          example:
            title:
            - Title cannot be empty
        message:
          type: string
          example: Invalid request
        retry_at:
          type:
          - string
          - 'null'
          format: date-time
      example:
        errors:
          title:
          - Title cannot be empty
        message: Invalid request
...
    MessageResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          example: Confirmation email sent successfully
      example:
        message: Confirmation email sent successfully
...
    StravaAuthorizeResponse:
      type: object
      required:
      - authorization_url
      properties:
        authorization_url:
          type: string
          example: https://www.strava.com/oauth/authorize?client_id=12345&response_type=code
      example:
        authorization_url: https://www.strava.com/oauth/authorize?client_id=12345&response_type=code
    StravaConnectionResponse:
      type: object
      required:
      - configured
      - connected
      - scopes
      - last_sync_status
      - last_sync_imported_count
      - last_sync_duplicate_count
      - last_sync_failed_count
      properties:
        athlete_id:
          type:
          - integer
          - 'null'
          format: int64
          example: 12345678
        athlete_name:
          type:
          - string
          - 'null'
          example: Alex Rider
        athlete_profile_medium_url:
          type:
          - string
          - 'null'
        athlete_username:
          type:
          - string
          - 'null'
          example: alex-rider
        configured:
          type: boolean
          example: true
        connected:
          type: boolean
          example: true
        last_sync_duplicate_count:
          type: integer
          format: int32
          example: 1
        last_sync_failed_count:
          type: integer
          format: int32
          example: 0
        last_sync_finished_at:
          type:
          - string
          - 'null'
          format: date-time
          example: 2026-09-27T12:00:00Z
        last_sync_imported_count:
          type: integer
          format: int32
          example: 3
        last_sync_message:
          type:
          - string
          - 'null'
        last_sync_started_at:
          type:
          - string
          - 'null'
          format: date-time
        last_sync_status:
          type: string
          example: succeeded
        last_synced_activity_started_at:
          type:
          - string
          - 'null'
          format: date-time
        scopes:
          type: array
          items:
            type: string
          example:
          - read
          - activity:read_all
      example:
        athlete_id: 12345678
        athlete_name: Alex Rider
        athlete_username: alex-rider
        configured: true
        connected: true
        last_sync_duplicate_count: 1
        last_sync_failed_count: 0
        last_sync_finished_at: 2026-09-27T12:00:00Z
        last_sync_imported_count: 3
        last_sync_status: succeeded
        scopes:
        - read
        - activity:read_all
    StravaWebhookChallengeResponse:
      type: object
      required:
      - hub.challenge
      properties:
        hub.challenge:
          type: string
          example: strava-verification-challenge
      example:
        hub.challenge: strava-verification-challenge
    StravaWebhookEvent:
      type: object
      required:
      - aspect_type
      - event_time
      - object_id
      - object_type
      - owner_id
      - subscription_id
      properties:
        aspect_type:
          type: string
          example: create
        event_time:
          type: integer
          format: int64
          example: 1790500000
        object_id:
          type: integer
          format: int64
          example: 1234567890
        object_type:
          type: string
          example: activity
        owner_id:
          type: integer
          format: int64
          example: 12345678
        subscription_id:
          type: integer
          format: int64
          example: 12345
        updates: {}
      example:
        aspect_type: create
        event_time: 1790500000
        object_id: 1234567890
        object_type: activity
        owner_id: 12345678
        subscription_id: 12345
        updates: {}
...
    UserResponse:
      type: object
      required:
      - pid
      - name
      - email
      - is_admin
      - verified
      properties:
        email:
          type: string
          example: alex@example.com
        is_admin:
          type: boolean
          example: false
        name:
          type: string
          example: Alex Rider
        pid:
          type: string
          example: a1b2c3d4-0000-4000-8000-000000000012
        verified:
          type: boolean
          example: true
      example:
        email: alex@example.com
        is_admin: false
        name: Alex Rider
        pid: a1b2c3d4-0000-4000-8000-000000000012
        verified: true
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...